+++ src/router.ts
@@ -2,6 +2,7 @@ import { createRouter, createWebHistory } from 'vue-router'
 import SplashView from './views/SplashView.vue'
 import HomeView from './views/HomeView.vue'
 import CreateCharacterView from './views/CreateCharacterView.vue'
+import LevelUpView from './views/LevelUpView.vue'
 import RulesView from './views/RulesView.vue'
 import RuleReaderView from './views/RuleReaderView.vue'
 import SimulatorView from './views/SimulatorView.vue'
@@ -18,6 +19,7 @@ const router=createRouter({
     {path:'/welcome',name:'welcome',component:SplashView},
     {path:'/characters',name:'characters',component:HomeView},
     {path:'/characters/create',name:'create-character',component:CreateCharacterView},
+    {path:'/characters/:id/level-up',name:'level-up-character',component:LevelUpView},
     {path:'/rules',name:'rules',component:RulesView},
     {path:'/rules/read/:slug',name:'rule-reader',component:RuleReaderView},
     {path:'/simulator/encounters/:id',name:'encounter',component:EncounterView},